refactor to now file structure
