--- tests/postprocess/helper.py
@@ -1,5 +1,0 @@
-from textwrap import dedent
-
-
-def normalize(program: str) -> str:
-    return dedent(program).strip()
