+++ tests/processing/post/test_sort_imports.py
@@ -1,13 +1,13 @@
 from openqasm3.parser import parse
 from openqasm3.printer import dumps
 
-from app.postprocess.sort_imports import SortImports
-from tests.postprocess.helper import normalize
+from app.processing.post.sort_imports import SortImports
+from app.processing.utils import normalize_qasm_string
 
 
 def test_basic() -> None:
     """Basic test that covers a single example."""
-    before = normalize("""
+    before = normalize_qasm_string("""
     include "stdgates.inc";
     bit[2] c;
     qubit[4] _all_qubits;
@@ -20,7 +20,7 @@ def test_basic() -> None:
     include "qelib1.inc";
     ccx q[0], q[1], q[3];
     """)
-    target = normalize("""
+    target = normalize_qasm_string("""
     include "stdgates.inc";
     include "qelib1.inc";
     bit[2] c;
@@ -32,5 +32,5 @@ def test_basic() -> None:
     cx q[1], q[2];
     ccx q[0], q[1], q[3];
     """)
-    actual = normalize(dumps(SortImports().visit(parse(before))))
+    actual = normalize_qasm_string(dumps(SortImports().visit(parse(before))))
     assert target == actual
